+++ backend/canisters/community/CHANGELOG.md
@@ -6,6 +6,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [unreleased]
 
+### Changed
+
+- Include channels invited to in `explore_channels` response ([#8492](https://github.com/open-chat-labs/open-chat/pull/8492))
+
 ## [[2.0.1863](https://github.com/open-chat-labs/open-chat/releases/tag/v2.0.1863-community)] - 2025-08-12
 
 ### Changed